Exception handling

+++ src/main/java/com/charnoski/ecommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.charnoski.ecommerce.services;
 import com.charnoski.ecommerce.dto.ProductDTO;
 import com.charnoski.ecommerce.entities.Product;
 import com.charnoski.ecommerce.repositories.ProductRepository;
+import com.charnoski.ecommerce.services.exceptions.DatabaseException;
+import com.charnoski.ecommerce.services.exceptions.ResouceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResouceNotFoundException("Recurso não encontrado")
+        );
         return new ProductDTO(product);
     }
 
@@ -37,15 +44,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResouceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if  (!productRepository.existsById(id)) {
+            throw new ResouceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
